+++ permissionsx/settings.py
@@ -11,3 +11,4 @@ from django.conf import settings
 PERMISSIONSX_REDIRECT_URL = getattr(settings, 'PERMISSIONSX_REDIRECT_URL', settings.LOGIN_URL)
 PERMISSIONSX_LOGOUT_IF_DENIED = getattr(settings, 'PERMISSIONSX_LOGOUT_IF_DENIED', False)
 PERMISSIONSX_DEBUG = getattr(settings, 'PERMISSIONSX_DEBUG', False)
+PERMISSIONSX_CACHING = getattr(settings, 'PERMISSIONSX_CACHING', False)