+++ permissionsx/tests/tests.py
@@ -5,60 +5,17 @@ PermissionsX - Authorization for Django.
 :license:   BSD, see LICENSE for more details.
 
 """
+from django.core.urlresolvers import reverse
+from django.contrib.auth import login
 
-from django.contrib.auth import (
-    login,
-    logout,
-)
-from django.contrib.auth.models import User
-from django.test import (
-    TestCase,
-    RequestFactory,
-)
-from django.contrib.sessions.backends.db import SessionStore
-from django.contrib.sessions.middleware import SessionMiddleware
-
-from permissionsx.models import (
-    Arg,
-    P,
-    Permissions,
-)
+from permissionsx.tests.test_utils import UtilityTestCase
+from permissionsx.tests.test_utils import SettingsOverride
+from permissionsx.models import Arg
 from permissionsx.tests.permissions import *
-from permissionsx.tests.models import Profile
+from permissionsx.tests.models import TestObject
 
 
-class PermissionsDefinitionsTests(TestCase):
-
-    def setUp(self):
-        self.session_backend = SessionStore()
-        self.session_middleware = SessionMiddleware()
-        self.factory = RequestFactory()
-        self.user = self.create_user('user')
-        self.admin = self.create_user('admin', is_superuser=True)
-        self.staff = self.create_user('staff', is_staff=True)
-
-    def create_user(self, username, **attrs):
-        if attrs is None:
-            attrs = {}
-        user = User.objects.create_user(username, username + '@example.com', 'password')
-        user.profile = Profile.objects.create(user=user)
-        user.__dict__.update(**attrs)
-        user.backend = self.session_backend
-        user.save()
-        return user
-
-    def get_request_for_user(self, user):
-        request = self.factory.get('/')
-        request.user = self.user
-        self.session_middleware.process_request(request)
-        logout(request)
-        return request
-
-    def permissions_for_request(self, permissions, request):
-        if isinstance(permissions, Permissions):
-            return permissions.check_permissions(request)
-        else:
-            return permissions().check_permissions(request)
+class PermissionsDefinitionsTests(UtilityTestCase):
 
     def test_is_authenticated(self):
         request = self.get_request_for_user(self.user)
@@ -99,8 +56,12 @@ class PermissionsDefinitionsTests(TestCase):
         self.assertFalse(self.permissions_for_request(AndStaffSuperuserPermissions, request_staff))
         self.assertFalse(self.permissions_for_request(AndStaffSuperuserPermissions, request_admin))
         self.assertFalse(self.permissions_for_request(AndStaffSuperuserPermissions, request_user))
-        self.admin.is_staff = True
         self.staff.is_superuser = True
+        self.admin.is_staff = True
+        # NOTE: User must log in again if permissions changed - cleaning
+        #       cache for a specific user.
+        login(request_staff, self.staff)
+        login(request_admin, self.admin)
         self.assertTrue(self.permissions_for_request(AndStaffSuperuserPermissions, request_staff))
         self.assertTrue(self.permissions_for_request(AndStaffSuperuserPermissions, request_admin))
         self.assertFalse(self.permissions_for_request(AndStaffSuperuserPermissions, request_user))
@@ -113,6 +74,7 @@ class PermissionsDefinitionsTests(TestCase):
         self.assertFalse(self.permissions_for_request(IsPublicPermissions, request_admin))
         self.assertFalse(self.permissions_for_request(AuthenticatedPermissions, request_admin))
         self.user.get_profile().is_public = True
+        login(request_user, self.user)
         self.assertTrue(self.permissions_for_request(IsPublicPermissions, request_user))
         self.assertFalse(self.permissions_for_request(IsPublicPermissions, request_admin))
 
@@ -127,6 +89,8 @@ class PermissionsDefinitionsTests(TestCase):
         self.assertFalse(self.permissions_for_request(NegatePermissions, request_admin))
         self.user.get_profile().is_public = True
         self.admin.get_profile().is_public = True
+        login(request_user, self.user)
+        login(request_admin, self.admin)
         self.assertTrue(self.permissions_for_request(NegatePermissions, request_user))
         self.assertTrue(self.permissions_for_request(NegatePermissions, request_admin))
 
@@ -136,8 +100,10 @@ class PermissionsDefinitionsTests(TestCase):
         login(request_admin, self.admin)
         self.assertFalse(self.permissions_for_request(NestedPermissions, request_admin))
         self.admin.username = 'admin2'
+        login(request_admin, self.admin)
         self.assertFalse(self.permissions_for_request(NestedPermissions, request_admin))
         self.admin.is_staff = True
+        login(request_admin, self.admin)
         self.assertTrue(self.permissions_for_request(NestedPermissions, request_admin))
 
     def test_request_params(self):
@@ -148,40 +114,48 @@ class PermissionsDefinitionsTests(TestCase):
 
     def test_overrides(self):
         request_admin = self.get_request_for_user(self.admin)
-        login(request_admin, self.admin)
         self.permissions_for_request(OverrideIfFalsePermissions, request_admin)
-        self.assertEqual('Override returns False', request_admin.permissionsx_return_overrides[False]())
-        self.permissions_for_request(OverrideIfTruePermissions, request_admin)
-        self.assertEqual('Override returns True', request_admin.permissionsx_return_overrides[True]())
+        self.assertEqual(OVERRIDE_FALSE, request_admin.permissionsx_return_overrides[0]())
         self.permissions_for_request(OverrideIfTrueFalsePermissions, request_admin)
-        self.assertEqual('Override returns False', request_admin.permissionsx_return_overrides[False]())
-        self.assertEqual('Override returns True', request_admin.permissionsx_return_overrides[True]())
+        self.assertEqual(OVERRIDE_FALSE, request_admin.permissionsx_return_overrides[0]())
         self.permissions_for_request(NegatedOverrideIfTrueFalsePermissions, request_admin)
-        self.assertEqual('Override returns True', request_admin.permissionsx_return_overrides[True]())
-        self.assertEqual('Override returns False', request_admin.permissionsx_return_overrides[False]())
+        self.assertEqual(OVERRIDE_FALSE, request_admin.permissionsx_return_overrides[0]())
         self.permissions_for_request(NestedNegatedOverridePermissions, request_admin)
-        self.assertEqual('Override returns True', request_admin.permissionsx_return_overrides[True]())
-        self.assertEqual('Override returns False', request_admin.permissionsx_return_overrides[False]())
+        self.assertEqual(OVERRIDE_FALSE, request_admin.permissionsx_return_overrides[0]())
+        login(request_admin, self.admin)
+        self.permissions_for_request(OverrideIfTruePermissions, request_admin)
+        self.assertEqual(OVERRIDE_TRUE, request_admin.permissionsx_return_overrides[0]())
+        self.permissions_for_request(OverrideIfTrueFalsePermissions, request_admin)
+        self.assertEqual(OVERRIDE_TRUE, request_admin.permissionsx_return_overrides[0]())
+        self.permissions_for_request(NegatedOverrideIfTrueFalsePermissions, request_admin)
+        self.assertEqual(OVERRIDE_TRUE, request_admin.permissionsx_return_overrides[0]())
+        self.permissions_for_request(NestedNegatedOverridePermissions, request_admin)
+        self.assertEqual(OVERRIDE_TRUE, request_admin.permissionsx_return_overrides[0]())
 
     def test_oneliners(self):
         request = self.get_request_for_user(self.user)
         self.assertFalse(self.permissions_for_request(Permissions(P(user__is_authenticated=True) & P(user__username=request.user.username)), request))
         login(request, self.user)
         self.assertTrue(self.permissions_for_request(Permissions(P(user__is_authenticated=True) | P(user__username='someotheruser')), request))
-        self.assertTrue(self.permissions_for_request(Permissions(P(user__is_authenticated=True) | P(user__username='someotheruser')), request))
         self.permissions_for_request(Permissions(user_is_authenticated & P(user__is_superuser=True, if_false=if_true_override)), request)
-        self.assertEqual('Override returns True', request.permissionsx_return_overrides[False]())
-        self.permissions_for_request(Permissions(user_is_superuser | user_is_staff, if_false=if_true_override), request)
-        self.assertEqual('Override returns True', request.permissionsx_return_overrides[False]())
+        self.assertEqual(OVERRIDE_TRUE, request.permissionsx_return_overrides[0]())
+        self.permissions_for_request(Permissions(user_is_superuser | P(user__is_staff=True, if_false=if_false_override)), request)
+        self.assertEqual(OVERRIDE_FALSE, request.permissionsx_return_overrides[0]())
         permissions_tested = Permissions(
             P(user__is_authenticated=True) &
             P(user__is_staff=True, if_false=if_false_override) &
-            P(user__is_staff=False, if_true=if_true_override) &
             P(user__is_superuser=False)
         )
         self.permissions_for_request(permissions_tested, request)
-        self.assertEqual('Override returns True', request.permissionsx_return_overrides[True]())
-        self.assertEqual('Override returns False', request.permissionsx_return_overrides[False]())
+        self.assertEqual(OVERRIDE_FALSE, request.permissionsx_return_overrides[0]())
+        login(request, self.staff)
+        permissions_tested = Permissions(
+            P(user__is_authenticated=True) &
+            P(user__is_staff=True, if_true=if_true_override) &
+            P(user__is_superuser=False)
+        )
+        self.permissions_for_request(permissions_tested, request)
+        self.assertEqual(OVERRIDE_TRUE, request.permissionsx_return_overrides[0]())
 
     def test_request_arguments(self):
         request = self.get_request_for_user(self.user)
@@ -190,3 +164,101 @@ class PermissionsDefinitionsTests(TestCase):
         self.assertTrue(self.permissions_for_request(Permissions(P(user__get_profile__is_attached_to_user=Arg('user'))), request))
         self.assertFalse(self.permissions_for_request(Permissions(P(user__get_profile__is_attached_to_user=Arg('user2'))), request))
 
+    def test_nested_negated(self):
+        request_admin = self.get_request_for_user(self.admin)
+        self.assertFalse(self.permissions_for_request(NestedNegatedPermissions, request_admin))
+        login(request_admin, self.admin)
+        self.assertTrue(self.permissions_for_request(NestedNegatedPermissions, request_admin))
+
+    def test_children(self):
+        self.assertEqual(
+            '(&(&(~(&{\'user\': 1})),(|{\'user\': 2},(~(&{\'user\': 3})))))',
+            str(P(~P(user=1) & P(P(user=2) | ~P(user=3))))
+        )
+
+
+class PermissionsDjangoViews(UtilityTestCase):
+
+    def setUp(self):
+        super(PermissionsDjangoViews, self).setUp()
+        self.test_object = TestObject.objects.create(title='Test!')
+
+    def test_authenticated(self):
+        response = self.client.get(reverse('authenticated'), follow=True)
+        self.assertContains(response, 'Login')
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('authenticated'), follow=True)
+        self.assertContains(response, 'Passed')
+
+    def test_get_profile(self):
+        # NOTE: Uses OrStaffSuperuserPermissions == P(user__is_staff=True) | P(user__is_administrator=True)
+        response = self.client.get(reverse('get_profile'), follow=True)
+        self.assertContains(response, 'Login')
+        self.login(self.client, 'user')
+        self.assertContains(response, 'Login')
+        self.login(self.client, 'staff')
+        response = self.client.get(reverse('get_profile'), follow=True)
+        self.assertContains(response, 'Passed')
+        self.login(self.client, 'admin')
+        response = self.client.get(reverse('get_profile'), follow=True)
+        self.assertContains(response, 'Passed')
+
+    def test_settings_logout(self):
+        with SettingsOverride(PERMISSIONSX_LOGOUT_IF_DENIED=True):
+            self.login(self.client, 'user')
+            response = self.client.get(reverse('authenticated'), follow=True)
+            self.assertContains(response, 'Passed')
+            self.client.get(reverse('superuser'), follow=True)
+            self.assertNotIn('_auth_user_id', self.client.session)
+
+    def test_settings_redirect_url(self):
+        with SettingsOverride(PERMISSIONSX_REDIRECT_URL='/accounts/login2/'):
+            response = self.client.get(reverse('authenticated'), follow=True)
+            self.assertContains(response, 'Login2')
+
+    def test_response_class(self):
+        response = self.client.get(reverse('response_class'), follow=True)
+        self.assertContains(response, 'Access Denied')
+        self.assertContains(response, 'Login')
+
+    def test_if_false_override(self):
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('overrides_if_false'), follow=True)
+        self.assertContains(response, 'Login2')
+
+    def test_if_true_override(self):
+        response = self.client.get(reverse('overrides_if_true'), follow=True)
+        self.assertContains(response, 'Login')
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('overrides_if_true'), follow=True)
+        self.assertContains(response, 'Passed')
+        self.login(self.client, 'admin')
+        response = self.client.get(reverse('overrides_if_true'), follow=True)
+        self.assertContains(response, 'Welcome')
+
+    def test_both_overrides(self):
+        response = self.client.get(reverse('overrides_both'), follow=True)
+        self.assertContains(response, 'Login2')
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('overrides_both'), follow=True)
+        self.assertContains(response, 'Passed')
+        self.login(self.client, 'admin')
+        response = self.client.get(reverse('overrides_both'), follow=True)
+        self.assertContains(response, 'Welcome')
+
+    def test_subsequent_overrides(self):
+        response = self.client.get(reverse('subsequent_overrides'), follow=True)
+        self.assertContains(response, 'Login2')
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('subsequent_overrides'), follow=True)
+        self.assertContains(response, 'Welcome')
+
+    def test_template_tag(self):
+        response = self.client.get(reverse('menu'), follow=True)
+        self.assertContains(response, 'Login')
+        self.login(self.client, 'user')
+        response = self.client.get(reverse('menu'), follow=True)
+        self.assertContains(response, 'User Menu')
+        self.login(self.client, 'staff')
+        response = self.client.get(reverse('menu'), follow=True)
+        self.assertContains(response, 'Staff Menu')